+++ src/midi/index.ts
@@ -1,3 +1,4 @@
 export * from "./connection";
 export * from "./errors";
 export * from "./ports";
+export * from "./request-response";